Rename list module to list_projects
