+++ gvsbuild/list_projects.py
